开始游戏加载进度（未完成）

+++ src/main/java/org/bkl/game/LauncherCallback.java
@@ -1,7 +1,7 @@
 package org.bkl.game;
 
 public interface LauncherCallback {
-    void onProgress(String message, int progress);
+    void onProgress(String message, double progress);
     void onSuccess();
     void onError(String errorMessage);
     void onLog(String logMessage);
